requirements.txt
flask
onnxruntime
pillow
python-multipart
python-dotenv
flask-cors
ultralytics
opencv-python
numpy
